refactor: 이미지 핸들러 커스텀 훅
